requirements.txt
streamlit
pdfplumber
pymupdf
pytesseract
langchain
faiss-cpu
transformers
torch
pillow
sentence-transformers